% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\docType{data}
\name{applications}
\alias{applications}
\title{Assylum applications}
\format{A data frame with 12 variables: \code{year_month}, \code{applications},
\code{male}, \code{female}, \code{minors}, \code{unacc_minors}, \code{decisions},
\code{granted}, \code{rejected}, \code{dublin}, \code{others}, \code{avg_time}}
\usage{
applications
}
\description{
Monthly data from the Swedish Migration Agency.
}
\keyword{datasets}
